requirements.txt
kafka-python>=2.1.2,<3
pyspark==3.5.1
python-dotenv==1.0.1
duckdb>=1.3.2
numpy>=2.2,<3
pandas>=2.2.3,<3
